MusicSelectionManagerクラスを共用部Baseと特殊部に分離
