+++ Assets/MusicSelection/Scripts/MusicSelectionManager.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 7e3abfe87e22b194ea0e6da09ca53f4d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: cefc0d5f5de84135a906839cdaffd408
+timeCreated: 1723869222
